A simple linear scaling strategy.
A slightly more advanced scaling strategy, when BinaryScalingStrategy
isn't enough. LinearScalingStrategy#new takes two parameters: nuber of
workers, and per-worker job capacity. Total work is the sum of the
queued, scheduled, and retrying jobs. The strategy scales to the
minimum amount of workers to handle all of the jobs. No more than the
maximum amount of workers are ever scaled to. The default worker
capacity is 25, Sidekiq's default concurrency.